requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-cloud-firestore==2.14.0
google-generativeai==0.8.3
google-auth==2.23.4
cloudinary==1.36.0
requests==2.31.0
pydantic==2.10.6
langchain==0.3.17
deepgram-sdk==3.0.0
groq==0.4.1
gunicorn==21.2.0
packaging
